Move event_log.go to `y` package
It is going to need to be used from there and elsewhere, so move into
`y`
